requirements.txt
Biopython>=1.84
pandas==2.2
scikit-learn>=1.5
